Separate zipper into workspace and program.
The program segment is considered read-only, while the workspace is for intermediate computations. In this way, the program segment acts as a sort of "program counter" that explicitly controls execution while the workspace controls computation. Currently the workspace will handle expression reduction while the program section will handle statements.

This seems like a nice split, and also allows us to keep immutable program text, something Jared mentioned was important. I agree with Jared, because people cannot destroy program text when they trace programs by hand, and I find it difficult to modify the program text in my mind when I try to execute it in my head while looking at the source code.

I thought about doing this for Mini-SML earlier, but it was much harder to do without a clear statement-expression split like we have here. When would you "throw" your computation to the workspace versus keep it in the program text? Solving it for Mini-Python might make the Mini-SML answer more clear.

There's also the question of whether this config transformation could be derived from another form of semantics and also of whether it could be supported by a customization API.